+++ app_fenrir/src/main/kotlin/dev/ragnarok/fenrir/api/adapters/VideoAlbumDtoAdapter.kt
@@ -19,7 +19,7 @@ class VideoAlbumDtoAdapter : AbsAdapter<VKApiVideoAlbum>("VKApiVideoAlbum") {
         val album = VKApiVideoAlbum()
         val root = json.asJsonObject
         album.id = optInt(root, "id")
-        album.owner_id = optInt(root, "owner_id")
+        album.owner_id = optLong(root, "owner_id")
         album.title = optString(root, "title")
         album.count = optInt(root, "count")
         album.updated_time = optLong(root, "updated_time")
